Create vice president creation resolution.markdown (#49)
This is creating an agenda supplements folder and a markdown of the vice president creation resolution. This file will have a link contained in an upcoming agenda. All attachments to agendas should be handled in this manner so that the membership can access the information and be prepared for the meetings without having to take time from the meeting for everyone to read.