Minutes for 10/07/2016 (#61)
* Minutes for 10/07/2016

Last, but not least, the minutes for today.

* Update 2016-10-07-minutes.markdown

I put this in pretty hastily and didn't give myself enough time to look at how it was being rendered. On second look, I decided to add a lot more spacing and bullets to make it easier to read quickly.

* Update 2016-10-07-minutes.markdown

* Update 2016-10-07-minutes.markdown

My suggested changes?